13285: Extract `HealthCertificateExportCertificatesInfoViewModel` to a dedicated swift file.
